Break up raster.stats for consistency
This commit breaks up raster.stats into raster.op.stats
and raster.histogram
