+++ benchmark/src/main/scala/geotrellis/benchmark/WeightedAdd.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,7 +19,7 @@ package geotrellis.benchmark
 import geotrellis.engine._
 import geotrellis.raster._
 import geotrellis.engine.op.local._
-import geotrellis.raster.stats._
+import geotrellis.raster.op.stats._
 import geotrellis.raster.render._
 
 import com.google.caliper.Param
@@ -54,11 +54,11 @@ class WeightedAdd extends OperationBenchmark {
     val n = names.length
     val re = getRasterExtent(names(0), size, size)
 
-    source = 
+    source =
       (0 until n).map(i => RasterSource(names(i), re) * weights(i))
                  .reduce(_ + _)
 
-    sourceSeq = 
+    sourceSeq =
       (0 until n).map(i => RasterSource(names(i), re) * weights(i))
                  .localAdd
