+++ engine/src/main/scala/geotrellis/engine/package.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,7 +15,7 @@
  */
 package geotrellis
 
-import geotrellis.raster.stats.Histogram
+import geotrellis.raster.histogram.Histogram
 
 import geotrellis.raster._
 
@@ -50,9 +50,9 @@ package object engine {
     def flatMap[T](f: (A, B, C, D)=>Op[T]) = Op(f).apply(t._1, t._2, t._3, t._4)
   }
 
-  /** 
-   * Syntax for converting an iterable collection to 
-   * have methods to work with the results of those 
+  /**
+   * Syntax for converting an iterable collection to
+   * have methods to work with the results of those
    * operations executed in parallel
    */
   implicit class OpMapSeq[A](seq: Seq[Op[A]]) {
