+++ raster-test/src/test/scala/geotrellis/raster/render/GetColorBreaksSpec.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,13 +18,13 @@ package geotrellis.raster.render
 
 import geotrellis.raster._
 import geotrellis.vector.Extent
-import geotrellis.raster.stats._
+import geotrellis.raster.op.stats._
 
 import org.scalatest._
 
 import geotrellis.testkit._
 
-class GetColorBreaksSpec extends FunSpec 
+class GetColorBreaksSpec extends FunSpec
                             with TestEngine
                             with Matchers {
   describe("GetColorBreaks") {
