+++ raster/src/main/scala/geotrellis/raster/histogram/ArrayHistogram.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package geotrellis.raster.stats
+package geotrellis.raster.histogram
 
 import geotrellis.raster._
 import math.{abs, ceil, min, max, sqrt}
@@ -40,9 +40,9 @@ object ArrayHistogram {
 // TODO: can currently only handle non-negative integers
 
 /**
-  * Data object representing a histogram that uses an array for internal storage. 
+  * Data object representing a histogram that uses an array for internal storage.
   */
-class ArrayHistogram(val counts: Array[Int], var total: Int) 
+class ArrayHistogram(val counts: Array[Int], var total: Int)
     extends MutableHistogram {
   def size = counts.length
