+++ raster/src/main/scala/geotrellis/raster/histogram/ConstantHistogram.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package geotrellis.raster.stats
+package geotrellis.raster.histogram
 
 case class ConstantHistogram(value:Int,size:Int) extends Histogram {
   def copy = ConstantHistogram(value,size)
@@ -33,4 +33,3 @@ case class ConstantHistogram(value:Int,size:Int) extends Histogram {
     fmh
   }
 }
-
