+++ raster/src/main/scala/geotrellis/raster/histogram/Histogram.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package geotrellis.raster.stats
+package geotrellis.raster.histogram
 
+import geotrellis.raster.op.stats.Statistics
 import geotrellis.raster.NODATA
 import math.{abs, round, sqrt}
