+++ raster/src/main/scala/geotrellis/raster/op/stats/package.scala
@@ -1,0 +1,8 @@
+package geotrellis.raster.op
+
+import geotrellis.raster.Tile
+import geotrellis.raster.op.stats.StatsMethods
+
+package object stats {
+  implicit class StatsMethodExtensions(val tile: Tile) extends StatsMethods { }
+}
