+++ raster/src/main/scala/geotrellis/raster/op/zonal/ZonalHistograms.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,7 +17,7 @@
 package geotrellis.raster.op.zonal
 
 import geotrellis.raster._
-import geotrellis.raster.stats._
+import geotrellis.raster.histogram._
 
 import spire.syntax.cfor._
 
@@ -30,7 +30,7 @@ import scala.collection.mutable
  *          If you use a Raster with a Double CellType (TypeFloat, TypeDouble)
  *          the data values will be rounded to integers.
  */
-object ZonalHistogram { 
+object ZonalHistogram {
 
   def apply(tile: Tile, zones: Tile): Map[Int, Histogram] = {
     val histMap = mutable.Map[Int, FastMapHistogram]()
