+++ raster/src/main/scala/geotrellis/raster/op/zonal/ZonalMethods.scala
@@ -1,7 +1,7 @@
 package geotrellis.raster.op.zonal
 
 import geotrellis.raster._
-import geotrellis.raster.stats.Histogram
+import geotrellis.raster.histogram.Histogram
 
 trait ZonalMethods extends TileMethods {
   def zonalHistogram(zones: Tile): Map[Int, Histogram] =
