+++ raster/src/main/scala/geotrellis/raster/op/zonal/summary/Histogram.scala
@@ -3,7 +3,7 @@ package geotrellis.raster.op.zonal.summary
 import geotrellis.raster._
 import geotrellis.vector._
 import geotrellis.raster.rasterize._
-import geotrellis.raster.stats._
+import geotrellis.raster.histogram._
 
 object Histogram extends TileIntersectionHandler[Histogram, Histogram] {
   def handlePartialTile(pt: PartialTileIntersection): Histogram = {
