+++ raster/src/main/scala/geotrellis/raster/render/png/Renderer.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,12 +19,12 @@ package geotrellis.raster.render.png
 import geotrellis._
 import geotrellis.raster._
 import geotrellis.raster.render._
-import geotrellis.raster.stats.Histogram
+import geotrellis.raster.histogram.Histogram
 
 import scala.collection.mutable
 
 case class Renderer(colorMap: ColorMap, cellType: CellType, colorType: ColorType) {
-  def render(r: Tile) = 
+  def render(r: Tile) =
     colorMap.render(r).convert(cellType)
   def settings = Settings(colorType, PaethFilter)
 }
