--- raster/src/main/scala/geotrellis/raster/stats/package.scala
@@ -1,5 +1,0 @@
-package geotrellis.raster
-
-package object stats {
-  implicit class StatsMethodExtensions(val tile: Tile) extends StatsMethods { }
-}
