+++ spark/src/main/scala/geotrellis/spark/io/LayerMetaData.scala
@@ -1,6 +1,6 @@
 package geotrellis.spark.io
 
-import geotrellis.raster.stats.Histogram
+import geotrellis.raster.histogram.Histogram
 import geotrellis.spark.RasterMetaData
 
 /** This is a container that represents all that the catalog knows about a layer */
