+++ spark/src/test/scala/geotrellis/spark/io/accumulo/AccumuloAttributeCatalogSpec.scala
@@ -4,7 +4,7 @@ import java.io.IOException
 
 import geotrellis.raster._
 import geotrellis.raster.json._
-import geotrellis.raster.stats._
+import geotrellis.raster.histogram._
 
 import geotrellis.spark._
 
@@ -24,7 +24,7 @@ class AccumuloAttributeCatalogSpec extends FunSpec
   with OnlyIfCanRunSpark {
 
   describe("Accumulo Attribute Catalog") {
-    ifCanRunSpark { 
+    ifCanRunSpark {
 
       val accumulo = new AccumuloInstance(
         instanceName = "fake",
